requirements/production.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    manage/update-deps.sh
#
backports.functools-lru-cache==1.6.1  # via soupsieve
beautifulsoup4==4.9.1
certifi==2020.6.20        # via requests
cffi==1.14.2              # via cryptography
chardet==3.0.4            # via requests
cryptography==3.1         # via pyopenssl, requests
django-anymail[mailgun]==7.2.1
django-markdown-deux==1.0.5
django-postman==3.7.0
django-translatable==0.1.1
django==1.11.29
enum34==1.1.10            # via cryptography
idna==2.10                # via requests
ipaddress==1.0.23         # via cryptography
iso8601==0.1.12           # via pychal
markdown2==2.3.9
pillow==6.2.2             # via reportlab
psycopg2-binary==2.8.6
pychal==1.10.0
pycparser==2.20           # via cffi
pyopenssl==19.1.0         # via requests
python-dateutil==2.8.1
pytz==2019.3              # via django, pychal, tzlocal
reportlab==3.5.49
requests[security]==2.23.0
six==1.15.0               # via cryptography, django-anymail, pyopenssl, python-dateutil
soupsieve==1.9.6          # via beautifulsoup4
stripe==2.51.0
tzlocal==2.0.0            # via pychal
urllib3==1.25.10          # via requests
uwsgi==2.0.19.1
xlwt==1.3.0
